+++ include/ccli/ccli_string.h
@@ -22,11 +22,11 @@ namespace ccli
 		operator std::string() { return m_String; }
 
 		// Moves until first non-whitespace char, returns the index at that spot
-		std::pair<size_t, size_t> NextPoi(unsigned long &start) const
+		std::pair<size_t, size_t> NextPoi(size_t &start) const
 		{
 			size_t end = m_String.size();
 			std::pair<size_t, size_t> range(end + 1, end);
-			unsigned long pos = start;
+			size_t pos = start;
 			for (; pos < end; ++pos)
 				if (!std::isspace(m_String[pos]))
 				{
